Filtering data with the PyQt5 builting class is solw.
Since, using  (or subclassing) the QSortFilterProxyModel class
of Qt have many advantages, but to filter thousand or more than  a thousand
of row is very very slow.
This commit is just made to keep the trace of this performance issue in case the
Qt provide a better way to handle the filtering.